+++ code/src/cljs/sixsq/nuvla/ui/edges_detail/views.cljs
@@ -21,6 +21,8 @@
             [sixsq.nuvla.ui.plugins.events :as events-plugin]
             [sixsq.nuvla.ui.plugins.tab :as tab-plugin]
             [sixsq.nuvla.ui.resource-log.views :as log-views]
+            [sixsq.nuvla.ui.routing.subs :as route-subs]
+            [sixsq.nuvla.ui.routing.utils :refer [name->href]]
             [sixsq.nuvla.ui.session.subs :as session-subs]
             [sixsq.nuvla.ui.utils.general :as general-utils]
             [sixsq.nuvla.ui.utils.map :as map]
@@ -30,7 +32,8 @@
             [sixsq.nuvla.ui.utils.time :as time]
             [sixsq.nuvla.ui.utils.ui-callback :as ui-callback]
             [sixsq.nuvla.ui.utils.values :as values]
-            [sixsq.nuvla.ui.utils.view-components :refer [OnlineStatusIcon]]))
+            [sixsq.nuvla.ui.utils.view-components :refer [OnlineStatusIcon]]
+            [sixsq.nuvla.ui.routing.routes :as routes]))
 
 
 (def refresh-action-id :nuvlabox-get-nuvlabox)
@@ -1992,17 +1995,36 @@
                      :edit-event      ::events/edit})])))
 
 
-(defn TabsNuvlaBox
+;; TODO: Extract the view logic to use on other pages, e.g. deployments details
+(defn TabsNuvlaEdge
   []
-  [tab-plugin/Tab
-   {:db-path [::spec/tab]
-    :menu    {:secondary true
-              :pointing  true
-              :style     {:display        "flex"
-                          :flex-direction "row"
-                          :flex-wrap      "wrap"}}
-    :panes   (tabs)}])
-
+  (let [tab-views      (tabs)
+        menu-items     (map :menuItem tab-views)
+        content-by-key (zipmap (map #(get-in % [:menuItem :key]) tab-views)
+                               (map :render tab-views))
+        route          @(subscribe [::route-subs/current-route])
+        path-params    (:path-params route)
+        query-params   (:query-params route)
+        cur-view       (keyword (:view query-params))]
+                         ;; FIXME: extract the relevant CSS here from Semantic UI.
+    [:div [:div {:class ["ui pointing secondary menu"]
+                 :style {:display       "flex"
+                         :flex-direction "row"
+                         :flex-wrap      "wrap"}}
+           (for [menu-item menu-items
+                 :let [{:keys [content icon key]} menu-item]]
+             ^{:key key}
+             (let [href (if (= key cur-view)
+                          nil
+                          (name->href routes/edges-details
+                            path-params
+                            (assoc query-params :view key)))]
+               [:a {:class ["item" (when (= key cur-view) "active")]
+                    :href href}
+                [ui/Icon {:name icon}]
+                content]))]
+     [:div {:class ["ui bottom attached segment active tab"]}
+      [(get content-by-key cur-view (:overview content-by-key))]]]))
 
 (defn PageHeader
   []
@@ -2017,6 +2039,7 @@
 (defn EdgeDetails
   [uuid]
   (refresh uuid)
+(js/console.error "uuid" uuid)
   (let [tr        @(subscribe [::i18n-subs/tr])
         nb-status @(subscribe [::subs/nuvlabox-status])]
     [components/LoadingPage {:dimmable? true}
@@ -2035,5 +2058,5 @@
          [ui/Message {:warning true
                       :icon    "warning sign"
                       :content (tr [:nuvlaedge-outdated-telemetry-warning])}])]
-      [TabsNuvlaBox]
+      [TabsNuvlaEdge]
       [AddPlaybookModal]]]))